[Fix] 메인 시간 분기처리 및 코드개선 (#76)
